Validações e tratamentos atualizados

+++ src/controllers/pacientesController.js
@@ -22,10 +22,16 @@ const pacientesController = {
             }
         });
 
-        res.json("Paciente atualizado com sucesso!");
+        res.json(req.body);
     },
     deletarPaciente: async (req, res) => {
         const { id } = req.params; 
+
+        const pacienteExiste = await Pacientes.findByPk(id);
+
+        if (pacienteExiste == null) {
+            return res.status(404).json("Id não encontrado");
+        };
 
         await Pacientes.destroy({
             where: {
@@ -39,6 +45,13 @@ const pacientesController = {
 
     async criarPaciente(req, res) {
         const { nome, email, idade } = req.body;
+
+        const pacienteExiste = await Pacientes.findOne({ where: { email } });
+
+        if (pacienteExiste) {
+            return res.status(400).json("Email já cadastrado");
+        }
+
         const novoPaciente = await Pacientes.create({ nome, email, idade });
         return res.status(201).json(novoPaciente);
     },
